% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a nonlinear model using C++ (power law, L2 loss only)}
\usage{
fit_model(x, y, model_type = "power_law", start_params = c(1, 1))
}
\arguments{
\item{x}{Numeric predictor}

\item{y}{Numeric response}

\item{model_type}{Only "power_law" is supported for now}

\item{start_params}{Starting values for model parameters (default: c(1,1))}
}
\value{
S3 object of class 'smartFit'
}
\description{
Fit a nonlinear model using C++ (power law, L2 loss only)
}
